Fix: restructured the components to be more cleaner
