+++ src/v1.1/components/GlobalFooter/FooterContainer/footerContainer.sass
@@ -1,0 +1,39 @@
+@import '~styles/variables'
+@import '~styles/mixins'
+
+.footerContainer
+  &__wrap
+    +clearfix()
+  &__title
+    padding-bottom: 16px
+    margin-bottom: 10px
+    border-bottom: 2px solid #E2DD12
+  &__listItem
+    padding-top: 20px
+
+// XSmall
++render-to($media-xs)
+  .footerContainer
+    margin-bottom: 34px
+    &__title
+      position: relative
+      &::after
+        content: ''
+        display: block
+        width: 18px
+        height: 18px
+        position: absolute
+        top: 3px
+        right: 15px
+        background-image: url($footer-arrow-icon)
+        background-size: cover
+        background-repeat: no-repeat
+        background-position: center
+    &__context
+      display: none
+    &.active
+      .footerContainer__title
+        &::after
+          transform: rotate(180deg)
+      .footerContainer__context
+        display: block
