+++ src/v1.1/components/GlobalFooter/FooterContainer/index.js
@@ -1,0 +1,2 @@
+import FooterContainer from './FooterContainer'
+export default FooterContainer
