+++ src/v1.1/components/GlobalFooter/FooterItemTitle/FooterItemTitle.js
@@ -1,13 +1,25 @@
 import React from 'react'
+import classNames from 'classnames'
 import { Link } from 'react-static'
 
+import './footerItemTitle.sass'
+
 const FooterItemTitle = props => {
-  const { title, to, onClick } = props
+  const {
+    title,
+    to,
+    active,
+    onClick,
+  } = props
+
   if (!title) {
     return <span>{' '}</span>
   }
+
   return (
-    <div className="footItem__title navigationText" onClick={onClick}>
+    <div className={classNames('footerItemTitle navigationText', {
+      active: active
+    })} onClick={onClick}>
       {
         to ? (<Link to={to}><h3>{title}</h3></Link>) : (<h3>{title}</h3>)
       }
