+++ src/v1.1/components/GlobalFooter/FooterItemTitle/footerItemTitle.sass
@@ -1,0 +1,27 @@
+@import '~styles/variables'
+@import '~styles/mixins'
+
+.footerItemTitle
+  padding-bottom: 16px
+  margin-bottom: 10px
+  border-bottom: 2px solid #E2DD12
+
+// XSmall
++render-to($media-xs)
+  .footerItemTitle
+    position: relative
+    &::after
+      content: ''
+      display: block
+      width: 18px
+      height: 18px
+      position: absolute
+      top: 3px
+      right: 15px
+      background-image: url($footer-arrow-icon)
+      background-size: cover
+      background-repeat: no-repeat
+      background-position: center
+    &.active
+      &::after
+        transform: rotate(180deg)
