+++ src/v1.1/components/GlobalFooter/FooterItemTitle/index.js
@@ -1,0 +1,2 @@
+import FooterItemTitle from './FooterItemTitle'
+export default FooterItemTitle
