+++ src/v1.1/components/GlobalFooter/FooterList/FooterList.js
@@ -1,27 +1,29 @@
 import React from 'react'
 import { Link } from 'react-static'
 
-import FootContainer from './footer_container'
-import { genRandomKey } from '~src/utils'
+import FooterContainer from '../FooterContainer'
+import { genRandomKey } from 'src/utils'
+
+import './footerList.sass'
 
 const FooterListContainer = props =>  {
   const { data, id } = props
   return (
-    <FootContainer id={id} name={data.name} to={data.to}>
-      <div className="footItem__list">
+    <FooterContainer id={id} name={data.name} to={data.to}>
+      <div className="footerList">
         {
           (data.links && data.links.length !== 0) ? (
             data.links.map(link => (
               <div
                 key={genRandomKey()}
-                className="footItem__listItem navigationText"
+                className="footerList__item navigationText"
               >
                 <Link to={link.to}>{link.name}</Link>
               </div>
           ))) : null
         }
       </div>
-    </FootContainer>
+    </FooterContainer>
   )
 }
