+++ src/v1.1/components/GlobalFooter/FooterList/footerList.sass
@@ -1,0 +1,3 @@
+.footerList
+  &__item
+    padding-top: 20px
