+++ src/v1.1/components/GlobalFooter/FooterList/index.js
@@ -1,0 +1,2 @@
+import FooterList from './FooterList'
+export default FooterList
