+++ src/v1.1/components/GlobalFooter/FooterSocial/footerSocial.sass
@@ -1,0 +1,70 @@
+@import '~styles/variables'
+@import '~styles/mixins'
+@import '~styles/fonts'
+
+.footerSocial
+  &__wrap
+    display: flex
+    flex-wrap: wrap
+    margin-left: -40px
+  &__item
+    width: 50px
+    height: 50px
+    border-radius: 50%
+    margin: 30px 0 10px 40px
+    background-color: transparent
+    background-repeat: no-repeat
+    background-size: cover
+    background-position: center
+    a
+      display: block
+      width: 100%
+      height: 100%
+    .name
+      display: none
+      +font-1()
+      padding-top: 20px
+      font-size: 16px
+      letter-spacing: -0.02em
+      font-weight: normal
+  &__twitter
+    background-image: url($footer-twitter-icon)
+  &__facebook
+    background-image: url($footer-facebook-icon)
+  &__medium
+    background-image: url($footer-medium-icon)
+  &__youtube
+    background-image: url($footer-youtube-icon)
+  &__reddit
+    background-image: url($footer-reddit-icon)
+  &__telegram
+    background-image: url($footer-telegram-icon)
+  &__riot
+    background-image: url($footer-riot-icon)
+  &__github
+    background-image: url($footer-github-icon)
+
+// Small
++render-to($media-s)
+  .footerSocial
+    &__wrap
+      margin-left: -36px
+    &__item
+      margin-left: 36px
+      margin-top: 20px
+      width: 40px
+      height: 40px
+
+// XSmall
++render-to($media-xs)
+  .footerSocial
+    &__wrap
+      display: block
+      margin-left: 0
+    &__item
+      background-image: none !important
+      width: auto
+      height: auto
+      margin: 0
+      .name
+        display: block
