+++ src/v1.1/components/GlobalFooter/FooterSocial/index.js
@@ -1,0 +1,2 @@
+import FooterSocial from './FooterSocial'
+export default FooterSocial
