+++ src/v1.1/components/GlobalFooter/FooterSubscribe/index.js
@@ -1,0 +1,2 @@
+import FooterSubscribe from './FooterSubscribe'
+export default FooterSubscribe
