+++ src/v1.1/components/Subscribe/subscribe.sass
@@ -1,7 +1,7 @@
-@import '../../styles/variables'
-@import '../../styles/mixins'
-@import '../../styles/fonts'
-@import '../../styles/colors'
+@import '~styles/variables'
+@import '~styles/mixins'
+@import '~styles/fonts'
+@import '~styles/colors'
 
 .footSubscribe
   &__wrap
