+++ src/v1.1/styles/shared/typos.sass
@@ -1,5 +1,5 @@
-@import '../fonts'
-@import '../colors'
+@import '~styles/fonts'
+@import '~styles/colors'
 
 .buttonText
   +font-1()
