+++ src/v1.1/styles/variables.sass
@@ -24,7 +24,7 @@ $col-count-small: 6
 
 // Default paths
 $assetsPath: '../../../assets'
-$imagePath: '#{$assetsPath}/images'
+$imagePath: '~images'
 
 $menu-icon: '#{$imagePath}/common/mob_menu.svg'
 $menu-icon-dark: '#{$imagePath}/common/mob_menu_dark.svg'
